refactor: Store 클래스 삭제 - 기존 Store 클래스는 Lottos 객체를 생성하는 일종의 팩토리 클래스에   불과함으로써 Lottos 클래스가 해당 기능 담당해도 무관   (Store.pay -> Lottos.buyLottos) - Lottos.supplyLottos 메소드의 경우 테스트 용도로 남겨둠
