+++ src/main/java/lotto/controller/Controller.java
@@ -12,14 +12,14 @@ public class Controller {
         int countOfManualLotto = InputView.inputCountOfManualLotto(price);
         List<Lotto> manualLottos = InputView.inputManualLottos(countOfManualLotto);
 
-        Lottos lottos = Store.pay(manualLottos, price);
+        Lottos myLottos = Lottos.buyLottos(manualLottos, price);
 
-        ResultView.printCountOfLottos(Store.countOfLottos(price), countOfManualLotto);
-        ResultView.printResultOfPurchase(lottos.getLottos());
+        ResultView.printCountOfLottos(Lottos.countOfLottos(price), countOfManualLotto);
+        ResultView.printResultOfPurchase(myLottos.getLottos());
 
         LottoWinningCondition winningCondition = InputView.inputWinningConditionOfLotto();
 
-        LottoWinner winner = lottos.calculateWinner(winningCondition);
+        LottoWinner winner = myLottos.calculateWinner(winningCondition);
         float yield = winner.calculateYield(price);
 
         ResultView.printWinnerStats(winner);
