+++ src/main/java/lotto/domain/Lottos.java
@@ -1,11 +1,10 @@
 package lotto.domain;
 
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 public class Lottos {
+    public static final int LOTTO_PRICE_PER_ONE = 1000;
+
     private final List<Lotto> lottos;
 
     private Lottos(List<Lotto> lottos) {
@@ -14,6 +13,21 @@ public class Lottos {
 
     static Lottos supplyLottos(List<Lotto> lottos) {
         return new Lottos(lottos);
+    }
+
+    public static Lottos buyLottos(List<Lotto> manualLottos, int price) {
+        final List<Lotto> lottos = new ArrayList<>(manualLottos);
+        int count = countOfLottos(price) - manualLottos.size();
+        while (count > 0) {
+            lottos.add(LottoFactory.createLottoAutomatically());
+            count--;
+        }
+
+        return Lottos.supplyLottos(lottos);
+    }
+
+    public static int countOfLottos(int price) {
+        return price / LOTTO_PRICE_PER_ONE;
     }
 
     public List<Lotto> getLottos() {
